refactor: separate directories
