--- src/Station.js
@@ -1,5 +1,0 @@
-export default class Station {
-  constructor(name) {
-    this.name = name;
-  }
-}
