+++ src/managers/stationManager.js
@@ -1,6 +1,6 @@
-import { clearPage, createSubmitBtn, createTextInput, createTable } from './utils.js';
-import { getLocalStorage, setLocalStorage } from './storage.js';
-import { stationText as T } from './constants.js';
+import { clearPage, createSubmitBtn, createTextInput, createTable } from '../utils/utils.js';
+import { getLocalStorage, setLocalStorage } from '../utils/storage.js';
+import { stationText as T } from '../utils/constants.js';
 
 const app = document.getElementById('app');
 const STORAGE_KEY_STATION = 'stations';
